Resolved PR issues
