+++ components/[pageId]/DocumentPage/DocumentPage.tsx
@@ -575,7 +575,6 @@ function DocumentPageComponent({
           <ProposalStickyFooter
             page={page}
             proposal={proposal}
-            refreshProposal={refreshProposal}
             formAnswers={proposalFormFieldAnswers}
             isStructuredProposal={isStructuredProposal}
           />
