+++ components/proposals/ProposalPage/components/ProposalStickyFooter/ProposalStickyFooter.tsx
@@ -16,10 +16,8 @@ export function ProposalStickyFooter({
   proposal,
   formAnswers,
   page,
-  isStructuredProposal,
-  refreshProposal
+  isStructuredProposal
 }: {
-  refreshProposal: VoidFunction;
   proposal: ProposalWithUsersAndRubric;
   formAnswers: FieldAnswerInput[];
   page: { title: string; hasContent?: boolean; sourceTemplateId: string | null; type: PageType };
@@ -36,7 +34,6 @@ export function ProposalStickyFooter({
   async function onClick() {
     try {
       await publishProposal();
-      refreshProposal();
     } catch (error) {
       showMessage((error as Error).message, 'error');
     }
