+++ lib/proposals/__tests__/submitEvaluationResult.spec.ts
@@ -55,75 +55,6 @@ describe('submitEvaluationResult()', () => {
     expect(proposalEvaluationReview).toBeTruthy();
   });
 
-  it('Should throw error if user has already submitted evaluation result for pass_fail step with > 1 required reviews', async () => {
-    const { space, user } = await generateUserAndSpace({
-      isAdmin: false
-    });
-    const { id: proposalId } = await testUtilsProposals.generateProposal({
-      spaceId: space.id,
-      userId: user.id,
-      proposalStatus: 'published',
-      evaluationInputs: [
-        {
-          evaluationType: 'pass_fail',
-          title: 'Pass/Fail',
-          reviewers: [
-            {
-              group: 'user',
-              id: user.id
-            }
-          ],
-          requiredReviews: 3,
-          permissions: []
-        }
-      ]
-    });
-
-    const proposalEvaluations = await prisma.proposalEvaluation.findMany({
-      where: {
-        proposalId
-      }
-    });
-
-    const evaluation = proposalEvaluations[0];
-
-    await submitEvaluationResult({
-      decidedBy: user.id,
-      evaluation,
-      proposalId,
-      result: 'pass',
-      spaceId: space.id
-    });
-
-    const proposalEvaluationReview = await prisma.proposalEvaluationReview.findFirst({
-      where: {
-        evaluationId: evaluation.id
-      }
-    });
-
-    const updatedEvaluation = await prisma.proposalEvaluation.findUniqueOrThrow({
-      where: {
-        id: evaluation.id
-      },
-      select: {
-        result: true
-      }
-    });
-
-    expect(proposalEvaluationReview).toBeTruthy();
-    expect(updatedEvaluation.result).toBeNull();
-
-    await expect(
-      submitEvaluationResult({
-        decidedBy: user.id,
-        evaluation,
-        proposalId,
-        result: 'pass',
-        spaceId: space.id
-      })
-    ).rejects.toThrow(ActionNotPermittedError);
-  });
-
   it('Should update result after users have submitted evaluation result for pass_fail step with > 1 required reviews', async () => {
     const { space, user } = await generateUserAndSpace({
       isAdmin: false
