+++ package.json
@@ -95,7 +95,7 @@
     "@bangle.dev/utils": "^0.31.6",
     "@beam-australia/react-env": "^3.1.1",
     "@ceramicnetwork/http-client": "^5.6.0",
-    "@charmverse/core": "^0.47.1-rc-feat-proposal-m.6",
+    "@charmverse/core": "^0.47.1-rc-feat-proposal-m.7",
     "@column-resizer/core": "^1.0.2",
     "@composedb/devtools": "^0.7.1",
     "@composedb/devtools-node": "^0.7.1",
